Get multas por ID y pago de multas

+++ Backend/src/main/java/com/subastas/backend/repository/MultaRepository.java
@@ -2,11 +2,12 @@ package com.subastas.backend.repository;
 
 import com.subastas.backend.entity.Multa;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface MultaRepository extends JpaRepository<Multa, Long> {
+public interface MultaRepository extends JpaRepository<Multa, Integer> {
     List<Multa> findByUsuarioIdentificador(Integer idUsuario);
 }
